Django/requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3